Updated content
Under Job summary is not the summary of the job, agency modal is not linked to the agency modal. Should be linked to: https://usajobs.github.io/ATP-Support/job-announcement-playbook/details/summary/#learn-more-about-this-agency-link 

Removed bullet about information that has moved to another section. 

Fixed naming issues and linked to proper sections as required.  

Under Eligibility, Who May Apply should be linked to the same section of that name.  

It also should say: We have provided a place for agencies to communicate their mission and provide a link to additional information for job seekers to learn more about the agency. 

“Using this section will free up the job summary to explain what the job seeker would be doing in that job.” Should be linked to: https://usajobs.github.io/ATP-Support/job-announcement-playbook/details/summary/